Rename generators
