+++ src/components/canvas/Canvas.test.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import Canvas from './Canvas';
 
-jest.mock('../../generators/workspace/WorkspaceGenerator');
+jest.mock('../../generators/workspace/workspace-generator');
 
 describe('Canvas', () => {
   it('renders without crashing', () => {
